# docs/v1/openapi-v1.yaml
openapi: 3.1.0
info:
  title: ReadyForRobots Robot Customer Acquisition Intelligence API
  version: 1.0.0
  description: V1 contract for robot analysis, facility opportunity ranking, qualification, and outcomes.
servers:
  - url: /api/v1
paths:
  /robot-analyses:
    post:
      operationId: createRobotAnalysis
      summary: Start public robot product analysis
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRobotAnalysis'
      responses:
        '202':
          description: Analysis accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RobotAnalysisAccepted'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /robot-analyses/{analysisId}:
    get:
      operationId: getRobotAnalysis
      summary: Get analysis progress or draft profile
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Current analysis state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RobotAnalysis'
        '404':
          $ref: '#/components/responses/NotFound'
  /robot-analyses/{analysisId}/confirm:
    post:
      operationId: confirmRobotAnalysis
      summary: Confirm profile and append OEM corrections
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRobotProfile'
      responses:
        '200':
          description: Profile confirmed and search started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmedRobotProfile'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    analysisToken:
      type: apiKey
      in: cookie
      name: rfr_analysis_token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    AnalysisId:
      name: analysisId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateRobotAnalysis:
      type: object
      properties:
        source_url:
          type: [string, 'null']
          format: uri
        description:
          type: [string, 'null']
          maxLength: 4000
      anyOf:
        - required: [source_url]
        - required: [description]
    RobotAnalysisAccepted:
      type: object
      required: [schema_version, analysis_id, analysis_token, status, status_url]
      properties:
        schema_version:
          const: v1
        analysis_id:
          type: string
          format: uuid
        analysis_token:
          type: string
        status:
          const: queued
        status_url:
          type: string
    RobotAnalysis:
      type: object
      required: [schema_version, analysis_id, status, progress]
      properties:
        schema_version:
          const: v1
        analysis_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, crawling, extracting, needs_review, confirmed, finding_work, resolving_facilities, matching, ranking, complete, failed, unsupported_robot, insufficient_product_evidence]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        retryable:
          type: boolean
        warnings:
          type: array
          items:
            type: string
        profile_etag:
          type: [string, 'null']
        draft_profile:
          type: [object, 'null']
          additionalProperties: true
    ProfileCorrection:
      type: object
      required: [field_path, value, truth_state]
      properties:
        field_path:
          type: string
        value: {}
        truth_state:
          const: oem_verified
        note:
          type: [string, 'null']
          maxLength: 1000
    ConfirmRobotProfile:
      type: object
      required: [profile_etag, corrections]
      properties:
        profile_etag:
          type: string
        corrections:
          type: array
          items:
            $ref: '#/components/schemas/ProfileCorrection'
    ConfirmedRobotProfile:
      type: object
      required: [schema_version, robot_id, profile_version_id, opportunity_search_id]
      properties:
        schema_version:
          const: v1
        robot_id:
          type: integer
        profile_version_id:
          type: string
          format: uuid
        opportunity_search_id:
          type: string
          format: uuid
        critical_unknowns:
          type: array
          items:
            type: string
...
    ErrorEnvelope:
      type: object
      required: [schema_version, error]
      properties:
        schema_version:
          const: v1
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            field_errors:
              type: array
              items:
                type: object
                additionalProperties: true
            required_facts:
              type: array
              items:
                type: string
  responses:
    ValidationError:
      description: Request or evidence validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource version or transition conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'